Partially move matrix recalculation to GtkView
